server/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
alembic==1.12.1
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6
python-dotenv==1.0.0
apscheduler==3.10.4
cryptography==41.0.7
plaid-python==12.0.0
pandas==2.1.3
openpyxl==3.1.2
ofxtools==0.9.5
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
